+++ aws/logs_monitoring/lambda_function.py
@@ -3,54 +3,43 @@
 # This product includes software developed at Datadog (https://www.datadoghq.com/).
 # Copyright 2020 Datadog, Inc.
 
-import base64
-import gzip
 import json
 import os
-
 import boto3
-import botocore
-import itertools
 import re
-import urllib
 import logging
-from io import BytesIO, BufferedReader
-import requests
+logger = logging.getLogger()
+logger.setLevel(logging.getLevelName(os.environ.get("DD_LOG_LEVEL", "INFO").upper()))
 
+import requests
 from datadog_lambda.wrapper import datadog_lambda_wrapper
 from datadog_lambda.metric import lambda_stats
 from datadog import api
+
 from trace_forwarder.connection import TraceConnection
 from enhanced_lambda_metrics import (
     get_enriched_lambda_log_tags,
     parse_and_submit_enhanced_metrics,
 )
 from logs import forward_logs
+from parsing import parse
 from telemetry import (
     DD_FORWARDER_TELEMETRY_NAMESPACE_PREFIX,
     get_forwarder_telemetry_tags,
-    set_forwarder_telemetry_tags,
 )
 from settings import (
     DD_API_KEY,
     DD_FORWARD_LOG,
     DD_SKIP_SSL_VALIDATION,
-    DD_TAGS,
     DD_API_URL,
     DD_TRACE_INTAKE_URL,
-    DD_MULTILINE_LOG_REGEX_PATTERN,
     DD_SOURCE,
     DD_CUSTOM_TAGS,
     DD_SERVICE,
     DD_HOST,
     DD_FORWARDER_VERSION,
     DD_ADDITIONAL_TARGET_LAMBDAS,
-    DD_USE_VPC,
 )
-
-
-logger = logging.getLogger()
-logger.setLevel(logging.getLevelName(os.environ.get("DD_LOG_LEVEL", "INFO").upper()))
 
 
 # DD_API_KEY must be set
@@ -81,31 +70,9 @@ trace_connection = TraceConnection(
     DD_TRACE_INTAKE_URL, DD_API_KEY, DD_SKIP_SSL_VALIDATION
 )
 
-rds_regex = re.compile("/aws/rds/(instance|cluster)/(?P<host>[^/]+)/(?P<name>[^/]+)")
-
-cloudtrail_regex = re.compile(
-    "\d+_CloudTrail_\w{2}-\w{4,9}-\d_\d{8}T\d{4}Z.+.json.gz$", re.I
-)
-
 HOST_IDENTITY_REGEXP = re.compile(
     r"^arn:aws:sts::.*?:assumed-role\/(?P<role>.*?)/(?P<host>i-([0-9a-f]{8}|[0-9a-f]{17}))$"
 )
-
-
-if DD_MULTILINE_LOG_REGEX_PATTERN:
-    try:
-        multiline_regex = re.compile(
-            "[\n\r\f]+(?={})".format(DD_MULTILINE_LOG_REGEX_PATTERN)
-        )
-    except Exception:
-        raise Exception(
-            "could not compile multiline regex with pattern: {}".format(
-                DD_MULTILINE_LOG_REGEX_PATTERN
-            )
-        )
-    multiline_regex_start_pattern = re.compile(
-        "^{}".format(DD_MULTILINE_LOG_REGEX_PATTERN)
-    )
 
 def datadog_forwarder(event, context):
     """The actual lambda function entry point"""
@@ -327,500 +294,6 @@ def extract_host_from_cloudtrails(event):
                 match = HOST_IDENTITY_REGEXP.match(arn)
                 if match is not None:
                     event[DD_HOST] = match.group("host")
-
-def parse(event, context):
-    """Parse Lambda input to normalized events"""
-    metadata = generate_metadata(context)
-    event_type = "unknown"
-    try:
-        # Route to the corresponding parser
-        event_type = parse_event_type(event)
-        if logger.isEnabledFor(logging.DEBUG):
-            logger.debug(f"Parsed event type: {event_type}")
-        if event_type == "s3":
-            events = s3_handler(event, context, metadata)
-        elif event_type == "awslogs":
-            events = awslogs_handler(event, context, metadata)
-        elif event_type == "events":
-            events = cwevent_handler(event, metadata)
-        elif event_type == "sns":
-            events = sns_handler(event, metadata)
-        elif event_type == "kinesis":
-            events = kinesis_awslogs_handler(event, context, metadata)
-    except Exception as e:
-        # Logs through the socket the error
-        err_message = "Error parsing the object. Exception: {} for event {}".format(
-            str(e), event
-        )
-        events = [err_message]
-
-    set_forwarder_telemetry_tags(context, event_type)
-
-    return normalize_events(events, metadata)
-
-def generate_metadata(context):
-    metadata = {
-        "ddsourcecategory": "aws",
-        "aws": {
-            "function_version": context.function_version,
-            "invoked_function_arn": context.invoked_function_arn,
-        },
-    }
-    # Add custom tags here by adding new value with the following format "key1:value1, key2:value2"  - might be subject to modifications
-    dd_custom_tags_data = {
-        "forwardername": context.function_name.lower(),
-        "forwarder_memorysize": context.memory_limit_in_mb,
-        "forwarder_version": DD_FORWARDER_VERSION,
-    }
-
-    metadata[DD_CUSTOM_TAGS] = ",".join(
-        filter(
-            None,
-            [
-                DD_TAGS,
-                ",".join(
-                    ["{}:{}".format(k, v) for k, v in dd_custom_tags_data.items()]
-                ),
-            ],
-        )
-    )
-
-    return metadata
-
-def parse_event_type(event):
-    if "Records" in event and len(event["Records"]) > 0:
-        if "s3" in event["Records"][0]:
-            return "s3"
-        elif "Sns" in event["Records"][0]:
-            # it's not uncommon to fan out s3 notifications through SNS,
-            # should treat it as an s3 event rather than sns event.
-            sns_msg = event["Records"][0]["Sns"]["Message"]
-            try:
-                sns_msg_dict = json.loads(sns_msg)
-                if "Records" in sns_msg_dict and "s3" in sns_msg_dict["Records"][0]:
-                    return "s3"
-            except Exception:
-                if logger.isEnabledFor(logging.DEBUG):
-                    logger.debug(f"No s3 event detected from SNS message: {sns_msg}")
-            return "sns"
-        elif "kinesis" in event["Records"][0]:
-            return "kinesis"
-
-    elif "awslogs" in event:
-        return "awslogs"
-
-    elif "detail" in event:
-        return "events"
-    raise Exception("Event type not supported (see #Event supported section)")
-
-# Handle S3 events
-def s3_handler(event, context, metadata):
-    # Need to use path style to access s3 via VPC Endpoints
-    # https://github.com/gford1000-aws/lambda_s3_access_using_vpc_endpoint#boto3-specific-notes
-    if DD_USE_VPC:
-        s3 = boto3.client(
-            "s3",
-            os.environ["AWS_REGION"],
-            config=botocore.config.Config(s3={"addressing_style": "path"}),
-        )
-    else:
-        s3 = boto3.client("s3")
-    # if this is a S3 event carried in a SNS message, extract it and override the event
-    if "Sns" in event["Records"][0]:
-        event = json.loads(event["Records"][0]["Sns"]["Message"])
-
-    # Get the object from the event and show its content type
-    bucket = event["Records"][0]["s3"]["bucket"]["name"]
-    key = urllib.parse.unquote_plus(event["Records"][0]["s3"]["object"]["key"])
-
-    source = parse_event_source(event, key)
-    metadata[DD_SOURCE] = source
-    ##default service to source value
-    metadata[DD_SERVICE] = source
-    ##Get the ARN of the service and set it as the hostname
-    hostname = parse_service_arn(source, key, bucket, context)
-    if hostname:
-        metadata[DD_HOST] = hostname
-
-    # Extract the S3 object
-    response = s3.get_object(Bucket=bucket, Key=key)
-    body = response["Body"]
-    data = body.read()
-
-    # Decompress data that has a .gz extension or magic header http://www.onicos.com/staff/iz/formats/gzip.html
-    if key[-3:] == ".gz" or data[:2] == b"\x1f\x8b":
-        with gzip.GzipFile(fileobj=BytesIO(data)) as decompress_stream:
-            # Reading line by line avoid a bug where gzip would take a very long time (>5min) for
-            # file around 60MB gzipped
-            data = b"".join(BufferedReader(decompress_stream))
-
-    if is_cloudtrail(str(key)):
-        cloud_trail = json.loads(data)
-        for event in cloud_trail["Records"]:
-            # Create structured object and send it
-            structured_line = merge_dicts(
-                event, {"aws": {"s3": {"bucket": bucket, "key": key}}}
-            )
-            yield structured_line
-    else:
-        # Check if using multiline log regex pattern
-        # and determine whether line or pattern separated logs
-        data = data.decode("utf-8")
-        if DD_MULTILINE_LOG_REGEX_PATTERN and multiline_regex_start_pattern.match(data):
-            split_data = multiline_regex.split(data)
-        else:
-            split_data = data.splitlines()
-
-        # Send lines to Datadog
-        for line in split_data:
-            # Create structured object and send it
-            structured_line = {
-                "aws": {"s3": {"bucket": bucket, "key": key}},
-                "message": line,
-            }
-            yield structured_line
-
-def parse_event_source(event, key):
-    """Parse out the source that will be assigned to the log in Datadog
-    Args:
-        event (dict): The AWS-formatted log event that the forwarder was triggered with
-        key (string): The S3 object key if the event is from S3 or the CW Log Group if the event is from CW Logs
-    """
-    lowercase_key = str(key).lower()
-
-    # Determines if the key matches any known sources for Cloudwatch logs
-    if "awslogs" in event:
-        return find_cloudwatch_source(lowercase_key)
-
-    # Determines if the key matches any known sources for S3 logs
-    if "Records" in event and len(event["Records"]) > 0:
-        if "s3" in event["Records"][0]:
-            if is_cloudtrail(str(key)):
-                return "cloudtrail"
-
-            return find_s3_source(lowercase_key)
-
-    return "aws"
-
-def find_cloudwatch_source(log_group):
-    # e.g. /aws/rds/instance/my-mariadb/error
-    if log_group.startswith("/aws/rds"):
-        for engine in ["mariadb", "mysql"]:
-            if engine in log_group:
-                return engine
-        return "rds"
-
-    # e.g. Api-Gateway-Execution-Logs_xxxxxx/dev
-    if log_group.startswith("api-gateway"):
-        return "apigateway"
-
-    # e.g. dms-tasks-test-instance
-    if log_group.startswith("dms-tasks"):
-        return "dms"
-
-    # e.g. sns/us-east-1/123456779121/SnsTopicX
-    if log_group.startswith("sns/"):
-        return "sns"
-
-    for source in [
-        "/aws/lambda",  # e.g. /aws/lambda/helloDatadog
-        "/aws/codebuild",  # e.g. /aws/codebuild/my-project
-        "/aws/kinesis",  # e.g. /aws/kinesisfirehose/dev
-        "/aws/docdb",  # e.g. /aws/docdb/yourClusterName/profile
-        "/aws/eks",  # e.g. /aws/eks/yourClusterName/profile
-    ]:
-        if log_group.startswith(source):
-            return source.replace("/aws/", "")
-
-    # the below substrings must be in your log group to be detected
-    for source in [
-        "network-firewall",
-        "route53",
-        "vpc",
-        "fargate",
-        "cloudtrail",
-        "msk",
-        "elasticsearch",
-    ]:
-        if source in log_group:
-            return source
-
-    return "cloudwatch"
-
-def is_cloudtrail(key):
-    match = cloudtrail_regex.search(key)
-    return bool(match)
-
-def find_s3_source(key):
-    # e.g. AWSLogs/123456779121/elasticloadbalancing/us-east-1/2020/10/02/123456779121_elasticloadbalancing_us-east-1_app.alb.xxxxx.xx.xxx.xxx_x.log.gz
-    if "elasticloadbalancing" in key:
-        return "elb"
-
-    # e.g. AWSLogs/123456779121/vpcflowlogs/us-east-1/2020/10/02/123456779121_vpcflowlogs_us-east-1_fl-xxxxx.log.gz
-    if "vpcflowlogs" in key:
-        return "vpc"
-
-    # e.g. 2020/10/02/21/aws-waf-logs-testing-1-2020-10-02-21-25-30-x123x-x456x
-    if "aws-waf-logs" in key:
-        return "waf"
-
-    # e.g. AWSLogs/123456779121/redshift/us-east-1/2020/10/21/123456779121_redshift_us-east-1_mycluster_userlog_2020-10-21T18:01.gz
-    if "_redshift_" in key:
-        return "redshift"
-
-    # this substring must be in your target prefix to be detected
-    if "amazon_documentdb" in key:
-        return "docdb"
-
-    # the below substrings must be in your target prefix to be detected
-    for source in [
-        "amazon_codebuild",
-        "amazon_kinesis",
-        "amazon_dms",
-        "amazon_msk",
-        "network-firewall",
-        "cloudfront",
-    ]:
-        if source in key:
-            return source.replace("amazon_", "")
-
-    return "s3"
-
-def parse_service_arn(source, key, bucket, context):
-    if source == "elb":
-        # For ELB logs we parse the filename to extract parameters in order to rebuild the ARN
-        # 1. We extract the region from the filename
-        # 2. We extract the loadbalancer name and replace the "." by "/" to match the ARN format
-        # 3. We extract the id of the loadbalancer
-        # 4. We build the arn
-        idsplit = key.split("/")
-        # If there is a prefix on the S3 bucket, idsplit[1] will be "AWSLogs"
-        # Remove the prefix before splitting they key
-        if len(idsplit) > 1 and idsplit[1] == "AWSLogs":
-            idsplit = idsplit[1:]
-            keysplit = "/".join(idsplit).split("_")
-        # If no prefix, split the key
-        else:
-            keysplit = key.split("_")
-        if len(keysplit) > 3:
-            region = keysplit[2].lower()
-            name = keysplit[3]
-            elbname = name.replace(".", "/")
-            if len(idsplit) > 1:
-                idvalue = idsplit[1]
-                return "arn:aws:elasticloadbalancing:{}:{}:loadbalancer/{}".format(
-                    region, idvalue, elbname
-                )
-    if source == "s3":
-        # For S3 access logs we use the bucket name to rebuild the arn
-        if bucket:
-            return "arn:aws:s3:::{}".format(bucket)
-    if source == "cloudfront":
-        # For Cloudfront logs we need to get the account and distribution id from the lambda arn and the filename
-        # 1. We extract the cloudfront id  from the filename
-        # 2. We extract the AWS account id from the lambda arn
-        # 3. We build the arn
-        namesplit = key.split("/")
-        if len(namesplit) > 0:
-            filename = namesplit[len(namesplit) - 1]
-            # (distribution-ID.YYYY-MM-DD-HH.unique-ID.gz)
-            filenamesplit = filename.split(".")
-            if len(filenamesplit) > 3:
-                distributionID = filenamesplit[len(filenamesplit) - 4].lower()
-                arn = context.invoked_function_arn
-                arnsplit = arn.split(":")
-                if len(arnsplit) == 7:
-                    awsaccountID = arnsplit[4].lower()
-                    return "arn:aws:cloudfront::{}:distribution/{}".format(
-                        awsaccountID, distributionID
-                    )
-    if source == "redshift":
-        # For redshift logs we leverage the filename to extract the relevant information
-        # 1. We extract the region from the filename
-        # 2. We extract the account-id from the filename
-        # 3. We extract the name of the cluster
-        # 4. We build the arn: arn:aws:redshift:region:account-id:cluster:cluster-name
-        namesplit = key.split("/")
-        if len(namesplit) == 8:
-            region = namesplit[3].lower()
-            accountID = namesplit[1].lower()
-            filename = namesplit[7]
-            filesplit = filename.split("_")
-            if len(filesplit) == 6:
-                clustername = filesplit[3]
-                return "arn:aws:redshift:{}:{}:cluster:{}:".format(
-                    region, accountID, clustername
-                )
-    return
-
-# Handle CloudWatch logs
-def awslogs_handler(event, context, metadata):
-    # Get logs
-    with gzip.GzipFile(
-        fileobj=BytesIO(base64.b64decode(event["awslogs"]["data"]))
-    ) as decompress_stream:
-        # Reading line by line avoid a bug where gzip would take a very long
-        # time (>5min) for file around 60MB gzipped
-        data = b"".join(BufferedReader(decompress_stream))
-    logs = json.loads(data)
-
-    # Set the source on the logs
-    source = logs.get("logGroup", "cloudwatch")
-
-    # Use the logStream to identify if this is a CloudTrail event
-    # i.e. 123456779121_CloudTrail_us-east-1
-    if "_CloudTrail_" in logs["logStream"]:
-        source = "cloudtrail"
-    metadata[DD_SOURCE] = parse_event_source(event, source)
-
-    # Default service to source value
-    metadata[DD_SERVICE] = metadata[DD_SOURCE]
-
-    # Build aws attributes
-    aws_attributes = {
-        "aws": {
-            "awslogs": {
-                "logGroup": logs["logGroup"],
-                "logStream": logs["logStream"],
-                "owner": logs["owner"],
-            }
-        }
-    }
-
-    # Set host as log group where cloudwatch is source
-    if metadata[DD_SOURCE] == "cloudwatch" or metadata.get(DD_HOST, None) == None:
-        metadata[DD_HOST] = aws_attributes["aws"]["awslogs"]["logGroup"]
-
-    # When parsing rds logs, use the cloudwatch log group name to derive the
-    # rds instance name, and add the log name of the stream ingested
-    if metadata[DD_SOURCE] in ["rds", "mariadb", "mysql"]:
-        match = rds_regex.match(logs["logGroup"])
-        if match is not None:
-            metadata[DD_HOST] = match.group("host")
-            metadata[DD_CUSTOM_TAGS] = (
-                metadata[DD_CUSTOM_TAGS] + ",logname:" + match.group("name")
-            )
-            # We can intuit the sourcecategory in some cases
-            if match.group("name") == "postgresql":
-                metadata[DD_CUSTOM_TAGS] + ",sourcecategory:" + match.group("name")
-
-    # For Lambda logs we want to extract the function name,
-    # then rebuild the arn of the monitored lambda using that name.
-    # Start by splitting the log group to get the function name
-    if metadata[DD_SOURCE] == "lambda":
-        log_group_parts = logs["logGroup"].split("/lambda/")
-        if len(log_group_parts) > 1:
-            lowercase_function_name = log_group_parts[1].lower()
-            # Split the arn of the forwarder to extract the prefix
-            arn_parts = context.invoked_function_arn.split("function:")
-            if len(arn_parts) > 0:
-                arn_prefix = arn_parts[0]
-                # Rebuild the arn with the lowercased function name
-                lowercase_arn = arn_prefix + "function:" + lowercase_function_name
-                # Add the lowercased arn as a log attribute
-                arn_attributes = {"lambda": {"arn": lowercase_arn}}
-                aws_attributes = merge_dicts(aws_attributes, arn_attributes)
-
-                env_tag_exists = (
-                    metadata[DD_CUSTOM_TAGS].startswith("env:")
-                    or ",env:" in metadata[DD_CUSTOM_TAGS]
-                )
-                # If there is no env specified, default to env:none
-                if not env_tag_exists:
-                    metadata[DD_CUSTOM_TAGS] += ",env:none"
-
-    # The EKS log group contains various sources from the K8S control plane.
-    # In order to have these automatically trigger the correct pipelines they
-    # need to send their events with the correct log source.
-    if metadata[DD_SOURCE] == "eks":
-        if logs["logStream"].startswith("kube-apiserver-audit-"):
-            metadata[DD_SOURCE] = "kubernetes.audit"
-        elif logs["logStream"].startswith("kube-scheduler-"):
-            metadata[DD_SOURCE] = "kube_scheduler"
-        # In case the conditions above don't match we maintain eks as the source
-
-    # Create and send structured logs to Datadog
-    for log in logs["logEvents"]:
-        yield merge_dicts(log, aws_attributes)
-
-def merge_dicts(a, b, path=None):
-    if path is None:
-        path = []
-    for key in b:
-        if key in a:
-            if isinstance(a[key], dict) and isinstance(b[key], dict):
-                merge_dicts(a[key], b[key], path + [str(key)])
-            elif a[key] == b[key]:
-                pass  # same leaf value
-            else:
-                raise Exception(
-                    "Conflict while merging metadatas and the log entry at %s"
-                    % ".".join(path + [str(key)])
-                )
-        else:
-            a[key] = b[key]
-    return a
-
-
-# Handle Cloudwatch Events
-def cwevent_handler(event, metadata):
-    data = event
-
-    # Set the source on the log
-    source = data.get("source", "cloudwatch")
-    service = source.split(".")
-    if len(service) > 1:
-        metadata[DD_SOURCE] = service[1]
-    else:
-        metadata[DD_SOURCE] = "cloudwatch"
-    ##default service to source value
-    metadata[DD_SERVICE] = metadata[DD_SOURCE]
-
-    yield data
-
-
-# Handle Sns events
-def sns_handler(event, metadata):
-    data = event
-    # Set the source on the log
-    metadata[DD_SOURCE] = "sns"
-
-    for ev in data["Records"]:
-        # Create structured object and send it
-        structured_line = ev
-        yield structured_line
-
-# Handle CloudWatch logs from Kinesis
-def kinesis_awslogs_handler(event, context, metadata):
-    def reformat_record(record):
-        return {"awslogs": {"data": record["kinesis"]["data"]}}
-
-    return itertools.chain.from_iterable(
-        awslogs_handler(reformat_record(r), context, metadata) for r in event["Records"]
-    )
-
-def normalize_events(events, metadata):
-    normalized = []
-    events_counter = 0
-
-    for event in events:
-        events_counter += 1
-        if isinstance(event, dict):
-            normalized.append(merge_dicts(event, metadata))
-        elif isinstance(event, str):
-            normalized.append(merge_dicts({"message": event}, metadata))
-        else:
-            # drop this log
-            continue
-
-    """Submit count of total events"""
-    lambda_stats.distribution(
-        "{}.incoming_events".format(DD_FORWARDER_TELEMETRY_NAMESPACE_PREFIX),
-        events_counter,
-        tags=get_forwarder_telemetry_tags(),
-    )
-
-    return normalized
 
 def forward_metrics(metrics):
     """
