+++ aws/logs_monitoring/logs.py
@@ -36,7 +36,6 @@ from settings import (
 )
 
 logger = logging.getLogger()
-logger.setLevel(logging.getLevelName(os.environ.get("DD_LOG_LEVEL", "INFO").upper()))
 
 class RetriableException(Exception):
     pass
